+++ Tests/Functional/TestBundle/Controller/DefaultController.php
@@ -44,7 +44,8 @@ class DefaultController extends Controller
                     ['name' => 'Long Kenner', 'salary' => 4051.0],
                     ['name' => 'Kena Kea', 'salary' => 4090.0],
                     ['name' => 'Evita Chittum', 'salary' => 4639.0]
-                ]
+                ],
+                'kernelPath' => $this->get('kernel')->getRootDir()
             ]
         );
     }