+++ Tests/Twig/BasicTwigTest.php
@@ -349,6 +349,31 @@ class BasicTwigTest extends AbstractTwigTest
      *
      * @dataProvider formatProvider
      */
+    public function testDocumentTemplate($format)
+    {
+        try {
+            $document = $this->getDocument('documentTemplate.' . $format, $format);
+            static::assertNotNull($document, 'Document does not exist');
+
+            $sheet = $document->getSheet(0);
+            static::assertNotNull($sheet, 'Sheet does not exist');
+
+            static::assertEquals('Hello2', $sheet->getCell('A1')->getValue(), 'Unexpected value in A1');
+            static::assertEquals('World', $sheet->getCell('B1')->getValue(), 'Unexpected value in B1');
+            static::assertEquals('Foo', $sheet->getCell('A2')->getValue(), 'Unexpected value in A2');
+            static::assertEquals('Bar2', $sheet->getCell('B2')->getValue(), 'Unexpected value in B2');
+        } catch (Twig_Error_Runtime $e) {
+            static::fail($e->getMessage());
+        }
+    }
+
+    /**
+     * @param string $format
+     *
+     * @throws \PHPExcel_Exception
+     *
+     * @dataProvider formatProvider
+     */
     public function testRowIndex($format)
     {
         try {