+++ Tests/Twig/CsvTwigTest.php
@@ -28,7 +28,6 @@ class CsvTwigTest extends AbstractTwigTest
     // Tests
     //
 
-
     /**
      * @param string $format
      *
@@ -44,6 +43,27 @@ class CsvTwigTest extends AbstractTwigTest
             static::assertTrue(file_exists($path), 'File does not exist');
             static::assertGreaterThan(0, filesize($path), 'File is empty');
             static::assertEquals("\"Foo\",\"Bar\"".PHP_EOL."\"Hello\",\"World\"".PHP_EOL, file_get_contents($path), 'Unexpected content');
+
+        } catch (Twig_Error_Runtime $e) {
+            static::fail($e->getMessage());
+        }
+    }
+
+    /**
+     * @param string $format
+     *
+     * @throws \PHPExcel_Exception
+     *
+     * @dataProvider formatProvider
+     */
+    public function testDocumentTemplate($format)
+    {
+        try {
+            $path = $this->getDocument('documentTemplate.csv', $format);
+
+            static::assertTrue(file_exists($path), 'File does not exist');
+            static::assertGreaterThan(0, filesize($path), 'File is empty');
+            static::assertEquals("\"Hello2\",\"World\"".PHP_EOL."\"Foo\",\"Bar2\"".PHP_EOL, file_get_contents($path), 'Unexpected content');
 
         } catch (Twig_Error_Runtime $e) {
             static::fail($e->getMessage());