+++ Twig/Node/XlsDrawingNode.php
@@ -3,7 +3,6 @@
 namespace MewesK\TwigExcelBundle\Twig\Node;
 
 use Twig_Compiler;
-use Twig_Node;
 use Twig_Node_Expression;
 
 /**