+++ Wrapper/XlsRowWrapper.php
@@ -1,6 +1,7 @@
 <?php
 
 namespace MewesK\TwigExcelBundle\Wrapper;
+use Twig_Environment;
 
 /**
  * Class XlsSheetWrapper
@@ -14,17 +15,24 @@ class XlsRowWrapper extends AbstractWrapper
      */
     protected $context;
     /**
+     * @var Twig_Environment
+     */
+    protected $environment;
+    /**
      * @var XlsSheetWrapper
      */
     protected $sheetWrapper;
 
     /**
+     * XlsRowWrapper constructor.
      * @param array $context
+     * @param Twig_Environment $environment
      * @param XlsSheetWrapper $sheetWrapper
      */
-    public function __construct(array $context, XlsSheetWrapper $sheetWrapper)
+    public function __construct(array $context, Twig_Environment $environment, XlsSheetWrapper $sheetWrapper)
     {
         $this->context = $context;
+        $this->environment = $environment;
         $this->sheetWrapper = $sheetWrapper;
     }